requirements.txt
Django==1.7.7
django-allauth==0.19.1
django-cors-headers==1.0.0
django-rest-auth==0.3.4
djangorestframework==3.1.0
jsonfield==1.0.3
MySQL-python==1.2.5
oauthlib==0.7.2
python-openid==2.2.5
requests==2.6.0
requests-oauthlib==0.4.2
six==1.9.0
